requirements.txt
pandas>=2.0
requests>=2.31
python-dotenv>=1.0
tabulate>=0.9
